+++ src/test/converter2/issues/gh1973/a.ts
@@ -6,4 +6,4 @@
 /**
  * Comment for a
  */
-export const a = "a";
+export const a = () => "a";