+++ src/test/converter2/issues/gh1973/b.ts
@@ -6,4 +6,4 @@
 /**
  * Comment for b
  */
-export const b = "b";
+export const b = () => "b";